Added timer, basic gameplay, implemented dictionary api calls to see if words are in the dictionary.
FEATURE: Words entered are now added to the list ONLY if they are in the dictionary. (Based on what comes back from the API calls.)
FEATURE: A three minute countdown timer starts upon arriving to the gameplay activity.
FIX: Fixed a bug where it was possible to enter words more than once and they'd show up in the list to be scored.